Add `instances` to `uri.v1` & `req.v1`
These new procedures return the URI, or make a request to said URI, in
order to get a list of instances registered at invidious.io. This should
be a good first step towards giving users an easy way to get the
available instances, and maybe in the future automating choosing an
instance.